+++ Changelog.md
@@ -11,6 +11,7 @@
 - Add expressions: listmap[list, expression], listreduce[list, expression], listfilter[list, expression]
 - Add expression: exec[string]
 - Add expression: listfromrange[start, end]
+- Add expression: listsort[list list, string expression]
 
 ## 0.2.5 (2021-10-06) [UtilityBeltInstaller-0.2.5.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/dd7285b144ef9327bf8041c64841c22f/UtilityBeltInstaller-0.2.5.exe)
 - Add expressions: wobjectgetdoubleprop[], wobjectgetboolprop[], wobjectgetstringprop[], wobjecthasdata[], wobjectrequestdata[]